project-initialization
Controlling the character by adding physical strength to it has been implemented. Added logical and physical behavior of the environment: friction, resistance, attraction. When jumping, the normal of the object and the vertical vector are compared, if the difference in angles is more than 45 degrees, the player cannot push off from such an object.